requirements.txt
google_api_python_client==1.6.1
tinydb==3.2.2
jsondiff==1.1.1
termcolor==1.1.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pyyaml==3.12